refactor: move text-area core and base CSS to styles folder
