+++ packages/text-area/src/styles/vaadin-text-area-base-styles.d.ts
