+++ packages/text-area/src/styles/vaadin-text-area-core-styles.js
